requirements.txt
bcrypt
fastapi
mysql-connector-python